refresh 2.2

+++ jogodeadivinhacao.c
@@ -9,7 +9,7 @@ int main() {
     int numerosecreto = 42;
     int chute;
     int tentativas = 1;
-    int pontos = 1000;
+    double pontos = 1000;
 
     while (1) {
         printf("Tentativa %d\n", tentativas);
@@ -33,10 +33,10 @@ int main() {
         }
         tentativas = tentativas + 1;
 
-        int pontosperdidodos = (chute - numerosecreto) /2;
+        double pontosperdidodos = (chute - numerosecreto) /2.0;
         pontos = pontos - pontosperdidodos;
     }
     printf ("Fim de Jogo!\n");
     printf ("Você acertou em %d tentativas\n!", tentativas);
-    printf ("Total de pontos: %d\n", pontos);
+    printf ("Total de pontos: %f\n", pontos);
 }
